Temp photo files are now deleted after encryption
I can't remember if I had photo/video importation working before but I do now.
I had to save the file from the photo picker in the app's directory to access
the data. I can improve things by importing the file to a location not viewable
in the files app and clearing the contents at various points in the app
lifecyle.